PWN-8348 - changes according to comments
